+++ packages/go_router/CHANGELOG.md
@@ -7,6 +7,14 @@
   describe the meaning of the child argument in the builder functions.
 - Added support for restorationId to ShellRoute (and StatefulShellRoute).
 
+## 5.2.4
+
+- Fixes crashes when using async redirect.
+
+## 5.2.3
+
+- Fixes link for router configuration and sub-routes
+
 ## 5.2.2
 
 - Fixes `pop` and `push` to update urls correctly.